refactor: Move lib/ directory into src/ for better organization
- Move lib/url-parser.ts to src/lib/url-parser.ts
- Update import path in src/index.ts (../lib -> ./lib)
- Update import paths in test scripts (../lib -> ../src/lib)
- Remove empty lib/ directory from root
- Improves project structure consistency (all source code in src/)
- Simplifies import paths within src/ directory

Tested:
- ✅ All 209 tests passing
- ✅ Wrangler dev compiles successfully
- ✅ Formatter applied
